requirements.txt
appdirs==1.4.4
argcomplete==1.8.1
attrs==21.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brotli==1.0.9
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2023.11.17
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.15.1
colcon-defaults==0.2.8
colcon-devtools==0.2.5
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.13
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
command-not-found==0.3
coverage==7.4.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
docutils==0.17.1
empy==3.3.4
exceptiongroup==1.2.0
filelock==3.13.1
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
fsspec==2023.12.2
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
identify==2.5.33
idna==3.6
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.3.2
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.3
matplotlib==3.5.1
mccabe==0.6.1
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
netifaces==0.11.0
networkx==3.2.1
nodeenv==1.8.0
notify2==0.3.1
numpy==1.21.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==21.3
Pillow==9.0.1
platformdirs==4.1.0
pluggy==0.13.0
ply==3.11
pre-commit==3.3.3
psutil==5.9.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyserial==3.5
pytest==7.4.0
pytest-cov==4.1.0
pytest-repeat==0.9.3
pytest-rerunfailures==13.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
roman==3.3
rosdistro-modules==0.9.0
rospkg-modules==1.5.0
RPi.GPIO==0.7.1
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
spidev==3.6
sympy==1.9
systemd-python==234
toml==0.10.2
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
triton==2.1.0
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.1.0
virtualenv==20.25.0
wadllib==1.3.6
webencodings==0.5.1
zipp==1.0.0
